Move utils under /api to ease module separation
