+++ aci/compose.go
@@ -29,7 +29,7 @@ import (
 	"github.com/docker/compose-cli/api/compose"
 	"github.com/docker/compose-cli/api/context/store"
 	"github.com/docker/compose-cli/api/errdefs"
-	"github.com/docker/compose-cli/utils/formatter"
+	"github.com/docker/compose-cli/api/utils/formatter"
 )
 
 type aciComposeService struct {
