+++ aci/convert/convert.go
@@ -34,7 +34,7 @@ import (
 	"github.com/docker/compose-cli/api/compose"
 	"github.com/docker/compose-cli/api/containers"
 	"github.com/docker/compose-cli/api/context/store"
-	"github.com/docker/compose-cli/utils/formatter"
+	"github.com/docker/compose-cli/api/utils/formatter"
 )
 
 const (
