+++ aci/e2e/e2e-aci_test.go
@@ -54,7 +54,7 @@ import (
 	"github.com/docker/compose-cli/api/context/store"
 	"github.com/docker/compose-cli/api/errdefs"
 	"github.com/docker/compose-cli/cli/cmd"
-	. "github.com/docker/compose-cli/utils/e2e"
+	. "github.com/docker/compose-cli/api/utils/e2e"
 )
 
 const (
