+++ cli/cmd/ps.go
@@ -29,7 +29,7 @@ import (
 	"github.com/docker/compose-cli/api/client"
 	"github.com/docker/compose-cli/api/containers"
 	formatter2 "github.com/docker/compose-cli/cli/formatter"
-	"github.com/docker/compose-cli/utils/formatter"
+	"github.com/docker/compose-cli/api/utils/formatter"
 )
 
 type psOpts struct {
