+++ cli/options/run/opts.go
@@ -22,7 +22,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/docker/compose-cli/utils"
+	"github.com/docker/compose-cli/api/utils"
 
 	"github.com/compose-spec/compose-go/types"
 	"github.com/containerd/containerd/platforms"
