+++ cli/server/proxy/containers.go
@@ -20,7 +20,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/docker/compose-cli/utils"
+	"github.com/docker/compose-cli/api/utils"
 
 	"github.com/compose-spec/compose-go/types"
 	"github.com/containerd/containerd/platforms"
