+++ ecs/e2e/ecs/e2e-ecs_test.go
@@ -31,7 +31,7 @@ import (
 	"gotest.tools/v3/icmd"
 	"gotest.tools/v3/poll"
 
-	. "github.com/docker/compose-cli/utils/e2e"
+	. "github.com/docker/compose-cli/api/utils/e2e"
 )
 
 var binDir string
